chapa_telegram_bot/requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0 